# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DirectorSaveRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "$ref": "#/components/schemas/DirectorBrief"
          },
          "input_fingerprint": {
            "anyOf": [
              {
                "pattern": "^[0-9a-f]{64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Fingerprint"
          },
          "mode": {
            "default": "manual",
            "enum": [
              "manual",
              "import",
              "ai"
            ],
            "title": "Mode",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt": {
            "default": "",
            "maxLength": 10000,
            "title": "Prompt",
            "type": "string"
          },
          "source_label": {
            "default": "",
            "maxLength": 500,
            "title": "Source Label",
            "type": "string"
          },
          "task_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "task_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Version"
          }
        },
        "required": [
          "content"
        ],
        "title": "DirectorSaveRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NarrativeApprovalRequest": {
        "additionalProperties": false,
        "properties": {
          "override_reason": {
            "default": "",
            "maxLength": 2000,
            "title": "Override Reason",
            "type": "string"
          }
        },
        "title": "NarrativeApprovalRequest",
        "type": "object"
      },
...
      "NarrativeFieldSuggestionRequest": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "default": "",
            "maxLength": 16000,
            "title": "Context",
            "type": "string"
          },
          "current_value": {
            "default": "",
            "maxLength": 8000,
            "title": "Current Value",
            "type": "string"
          },
          "field_key": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Field Key",
            "type": "string"
          },
          "field_label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Field Label",
            "type": "string"
          },
          "locale": {
            "default": "fr",
            "enum": [
              "fr",
              "en"
            ],
            "title": "Locale",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          }
        },
        "required": [
          "field_key",
          "field_label"
        ],
        "title": "NarrativeFieldSuggestionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "La Serre",
    "version": "0.2.13"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/narrative/field/suggest": {
      "post": {
        "operationId": "suggest_field_api_narrative_field_suggest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeFieldSuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Suggest Field Api Narrative Field Suggest Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest Field",
        "tags": [
          "narrative"
        ]
      }
    },
...
    "/api/narrative/series/director": {
      "put": {
        "operationId": "save_director_api_narrative_series_director_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectorSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Director Api Narrative Series Director Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Director",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/director/approve": {
      "post": {
        "operationId": "approve_director_api_narrative_series_director_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Director Api Narrative Series Director Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Director",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/director/generate": {
      "post": {
        "operationId": "generate_director_api_narrative_series_director_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Director Api Narrative Series Director Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Director",
        "tags": [
          "narrative"
        ]
      }
    },
...
    "/api/narrative/series/screenwriter": {
      "put": {
        "operationId": "save_screenwriter_api_narrative_series_screenwriter_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScreenwriterSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Screenwriter Api Narrative Series Screenwriter Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Screenwriter",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/screenwriter/approve": {
      "post": {
        "operationId": "approve_screenwriter_api_narrative_series_screenwriter_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Screenwriter Api Narrative Series Screenwriter Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Screenwriter",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/screenwriter/generate": {
      "post": {
        "operationId": "generate_screenwriter_api_narrative_series_screenwriter_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Screenwriter Api Narrative Series Screenwriter Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Screenwriter",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/validator": {
      "put": {
        "operationId": "save_validator_api_narrative_series_validator_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatorSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Save Validator Api Narrative Series Validator Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Validator",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/validator/approve": {
      "post": {
        "operationId": "approve_validator_api_narrative_series_validator_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Validator Api Narrative Series Validator Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Validator",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/series/validator/generate": {
      "post": {
        "operationId": "generate_validator_api_narrative_series_validator_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NarrativeGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Generate Validator Api Narrative Series Validator Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Validator",
        "tags": [
          "narrative"
        ]
      }
    },
    "/api/narrative/shot": {
      "post": {
        "operationId": "draft_shot_api_narrative_shot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShotDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Draft Shot Api Narrative Shot Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft Shot",
        "tags": [
          "narrative"
        ]
      }
    },
...
  }
}